Read ProjectSettings to pre-populate the logger class list to avoid needing to read and parse scripts.
I noticed my Autoloads were being labeled as (Node) and realized in the current system they could never not be (Node) since it's only looking for a class_name in the file. 

After adding autoloads, I found we could read the global_class_list to pre-populate class_names and in most cases, avoid needing FileAccess unless the class is not named.